requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
google-generativeai
chromadb
pytest
httpx
flake8